refactor: extract copy_button widget; use in update banner + welcome page
